+++ test/test_mate/cmd_test.clj
@@ -1,0 +1,7 @@
+(ns test-mate.cmd-test
+  (:require [midje.sweet   :refer :all]
+            [test-mate.cmd :refer :all]))
+
+(facts "exit"
+  (fact "nothing happens if exit called"
+    (exit-with-usage "message") => irrelevant)) ;ok if no exception thrown